Extra application
Move the current application to `app-service-1` and create a copy of
the code in `app-service-2`.

Nginx routes only to `app-service-1`.

The building of the applications occurs in `Dockerfile-app-service-1`
and `Dockerfile-app-service-2`. I think the building of the
application needs refactoring as it doesn't currently support
building applications that consist of multiple modules, plus the
build is slow as it does all of the gradle operations twice
(separately for each build), unless I can somehow add a cache to that
 is accessible to each build.