Use `kubernetes_ingress_v1` instead of `nginx`
Replace Nginx with Kubernetes Ingress.

Using Kubernetes Ingress removes the need for as much direct
configuration of Nginx which was found in the terraform configuration
before this commit. More precisely (from my perspective), the setup
of paths to other services is more dynamic and visible from the
image's configuration, rather than requiring direct configuration of
a Nginx config file.

The `external-dependencies` became `ingress` which contains the
configuration to setup `kubernetes_ingress_v1`. The module takes in a
map list of `path`, `service_name` and `service_port` combinations.